Refactored the data types detectors for more succinctness.
