+++ src/IsAFuzzyDataType.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of DataTypeValidator, a PHP Experts, Inc., Project.
  *
- * Copyright © 2019 PHP Experts, Inc.
+ * Copyright © 2019-2025 PHP Experts, Inc.
  * Author: Theodore R. Smith <theodore@phpexperts.pro>
  *  GPG Fingerprint: 4BF8 2613 1C34 87AC D28F  2AD8 EB24 A91D D612 5690
  *  https://www.phpexperts.pro/
@@ -14,7 +14,7 @@
 
 namespace PHPExperts\DataTypeValidator;
 
-class IsAFuzzyDataType extends IsAStrictDataType
+class IsAFuzzyDataType  extends IsADataType implements IsA
 {
     public function isBool($value): bool
     {
